Create README.md
This project consists in calculate the inverse of a matrix. For this we have two functions: makeCacheMatrix and cacheSolve. 
The first function calculate the inverse of a maytrix a cache it. The second function, first ask if the inverse of this matrix is in the cache, 
if the answer is yes, then get the matrix from the cache and return it, and if the answer is no, calculate the matrix.